Update to README
Just adding more content to the README.
I think an important part of making this successful will be making it
so people can understand how to use it.
I think that's doubly true when people aren't familiar with how
this protocol works, and so it's not like they know exactly what they
want to do with this library right off the bat.